refactor(session): split snapshot.go → snapshot.go + snapshot_send.go
snapshot.go was 235 lines after M-7 rewrite — split into:
- snapshot.go (133 lines): fields, persist, flush, delete lifecycle
- snapshot_send.go (111 lines): SendSnapshot, sendSnapshotFromCache,
  sendSnapshotForActor (all send-path functions)

Both files under 200-line limit.
